WIP: we must reduce context size.
Right now, all the source files are concatenated
with the README file to form the content that is
sent to the inference endpoint. We don't need that.

Solution: extract the imports from the source files
using pip freeze, generating a pre-requirements file
that should be compared to the requirements file
extracted from analyzing the README file.